Refactor ProductData to ExternalServices
